feat(frontier): implement Causal Frontier V2.0
- FrontierConfig struct with weak/medium/powerful presets
- FrontierEntity enum (Token(u32) | Connection(u32)) replaces separate pop_token/pop_connection
- Unified pop() with built-in causal budget check; tokens prioritized over connections
- begin_cycle()/end_cycle() replaces increment_processed()/reset_cycle()
- frontier_growth_rate and StormMetrics snapshot
- FrontierState::Stabilizing (renamed from Stabilized)
- EntityQueue: BitVec for O(1) deduplication, u32 indices throughout
- FrontierProcessor updated: EvaluationResult::Transform carries Vec<FrontierEntity>
- axiom-domain, axiom-heartbeat updated to V2.0 API
- 372 tests passing (+6 frontier/processor tests)

+++ crates/axiom-heartbeat/src/lib.rs
@@ -218,7 +218,7 @@ pub fn handle_heartbeat(
     if config.batch_size > 0 && total_tokens > 0 {
         for i in 0..config.batch_size {
             let token_idx = ((pulse_number as usize) * config.batch_size + i) % total_tokens;
-            frontier.push_token(token_idx);
+            frontier.push_token(token_idx as u32);
         }
     }
 
@@ -226,7 +226,7 @@ pub fn handle_heartbeat(
     if config.enable_connection_maintenance && config.connection_batch_size > 0 && total_connections > 0 {
         for i in 0..config.connection_batch_size {
             let conn_idx = ((pulse_number as usize) * config.connection_batch_size + i) % total_connections;
-            frontier.push_connection(conn_idx);
+            frontier.push_connection(conn_idx as u32);
         }
     }
 }
